Add DictToTable and ListToTable to tabledata
These are the same as the JSON modules, but take live Python objects
instead of a JSON file.
