+++ vistrails/packages/tabledata/read/__init__.py
@@ -37,6 +37,8 @@ from __future__ import division
 
 from vistrails.core.modules.utils import make_modules_dict
 
+from .convert import _modules as convert_modules
+
 try:
     # read_numpy requires numpy
     import numpy
@@ -54,6 +56,6 @@ else: # pragma: no cover
 from .read_csv import _modules as csv_modules
 from .read_json import _modules as json_modules
 
-_modules = make_modules_dict(numpy_modules, csv_modules, excel_modules,
-                             json_modules,
+_modules = make_modules_dict(convert_modules, numpy_modules, csv_modules,
+                             excel_modules, json_modules,
                              namespace='read')
