+++ vistrails/packages/tabledata/read/read_json.py
@@ -40,68 +40,19 @@ try:
 except ImportError:
     import json
 
-from vistrails.core.modules.vistrails_module import ModuleError
-
-from ..common import TableObject, Table, InternalModuleError
+from .convert import DictToTable, ListToTable
 
 
-class JSONTable(Table):
-    _input_ports = [('file', '(org.vistrails.vistrails.basic:File)')]
-    _output_ports = [
-            ('column_count', '(org.vistrails.vistrails.basic:Integer)'),
-            ('column_names', '(org.vistrails.vistrails.basic:List)'),
-            ('value', Table)]
-
+class JSONTable(object):
     def compute(self):
         json_file = self.get_input('file').name
         with open(json_file, 'rb') as fp:
             obj = json.load(fp)
-        try:
-            table = self.make_table(obj)
-        except InternalModuleError, e:
-            e.raise_module_error(self)
-        self.set_output('column_count', table.columns)
-        if table.names is not None:
-            self.set_output('column_names', table.names)
-        self.set_output('value', table)
 
-    @staticmethod
-    def add_list(columns, key, value):
-        if not isinstance(value, list):
-            raise InternalModuleError("Entry for key %r is not a list" % key)
-        if key is not None:
-            value = [key] + value
-        for i, v in enumerate(value):
-            columns[i].append(v)
-
-    @staticmethod
-    def add_dict(columns, keys, key_set, key, value):
-        if not isinstance(value, dict):
-            raise InternalModuleError(
-                    "Entry for key %r is not an object" % key)
-        value_keys = set(value.keys())
-        m = key_set - value_keys
-        if m:
-            raise InternalModuleError(
-                    "Entry for key %r is missing field %r" % (
-                    key,
-                    next(iter(m))))
-        m = value_keys - key_set
-        if m:
-            raise InternalModuleError(
-                    "Entry for key %r has unexpected field %r" % (
-                    key,
-                    next(iter(m))))
-        if key is None:
-            for i, k in enumerate(keys):
-                columns[i].append(value[k])
-        else:
-            columns[0].append(key)
-            for i, k in enumerate(keys):
-                columns[i + 1].append(value[k])
+        self.convert_to_table(obj)
 
 
-class JSONObject(JSONTable):
+class JSONObject(JSONTable, DictToTable):
     """Loads a JSON file and build a table from an object.
 
     In JSON, an object is written with {}. It is essentially an associative
@@ -125,45 +76,10 @@ class JSONObject(JSONTable):
 
     To read a list of rows, use the JSONList module instead.
     """
-    _input_ports = [('key_name', '(org.vistrails.vistrails.basic:String)',
-                     {'optional': True, 'defaults': repr(["key"])})]
-
-    def make_table(self, obj):
-        if not isinstance(obj, dict):
-            raise ModuleError(self, "JSON file is not an object")
-        key_name = self.get_input('key_name', True)
-        iterator = obj.iteritems()
-        try:
-            first_key, first_value = next(iterator)
-        except StopIteration:
-            raise ModuleError(self, "No entry in JSON object")
-        count = 1
-        if isinstance(first_value, list):
-            keys = None
-            columns = [[] for i in xrange(1 + len(first_value))]
-            self.add_list(columns, first_key, first_value)
-            for key, value in iterator:
-                self.add_list(columns, key, value)
-                count += 1
-        elif isinstance(first_value, dict):
-            keys = first_value.keys()
-            key_set = set(keys)
-            columns = [[] for i in xrange(1 + len(keys))]
-            self.add_dict(columns, keys, key_set, first_key, first_value)
-            for key, value in iterator:
-                self.add_dict(columns, keys, key_set, key, value)
-                count += 1
-        else:
-            raise ModuleError(self, "Values should be lists or objects")
-
-        if keys is not None:
-            names = [key_name] + keys
-        else:
-            names = None
-        return TableObject(columns, count, names)
+    _input_ports = [('file', '(org.vistrails.vistrails.basic:File)')]
 
 
-class JSONList(JSONTable):
+class JSONList(JSONTable, ListToTable):
     """Loads a JSON file and build a table from a list.
 
     In JSON, a list is written with [].
@@ -176,37 +92,10 @@ class JSONList(JSONTable):
 
         gives a 4x4 unnamed table.
     """
-    def make_table(self, obj):
-        if not isinstance(obj, list):
-            raise ModuleError(self, "JSON file is not a list")
-        iterator = iter(obj)
-        try:
-            first = next(iterator)
-        except StopIteration:
-            raise ModuleError(self, "No element in JSON list")
-        count = 1
-        if isinstance(first, list):
-            keys = None
-            columns = [[] for i in xrange(len(first))]
-            self.add_list(columns, None, first)
-            for value in iterator:
-                self.add_list(columns, None, value)
-                count += 1
-        elif isinstance(first, dict):
-            keys = first.keys()
-            key_set = set(keys)
-            columns = [[] for i in xrange(len(keys))]
-            self.add_dict(columns, keys, key_set, None, first)
-            for value in iterator:
-                self.add_dict(columns, keys, key_set, None, value)
-                count += 1
-        else:
-            raise ModuleError(self, "Values should be lists or objects")
-
-        return TableObject(columns, count, keys)
+    _input_ports = [('file', '(org.vistrails.vistrails.basic:File)')]
 
 
-_modules = [(JSONTable, {'abstract': True}), JSONObject, JSONList]
+_modules = [JSONObject, JSONList]
 
 
 ###############################################################################
